requirements.txt
streamlit
flake8
radon
black
python-dotenv
groq      # if AI API available
